requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.15
spacy>=3.5.0
networkx>=2.8.0
rank-bm25>=0.2.2
tenacity>=8.0.0
faiss-cpu>=1.7.0
tqdm>=4.65.0
numpy>=1.24.0
requests>=2.28.0
FlagEmbedding>=1.2.0
torch>=2.0.0 